app/requirements.txt
fastapi
uvicorn
requests
python-dotenv
flask
transformers
torch